Dual-Controler works now
