--- Twenty403/src/main/java/org/firstinspires/ftc/twenty403/Controls/ControlDriver.java
@@ -1,4 +1,0 @@
-package org.firstinspires.ftc.twenty403.Controls;
-
-public class ControlDriver {
-}
