+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/Controls/ControlsDriver.java
@@ -1,32 +1,19 @@
 package org.firstinspires.ftc.twenty403.Controls;
 
 import org.firstinspires.ftc.twenty403.Robot;
-import org.firstinspires.ftc.twenty403.command.claw.ClawCloseCommand;
-import org.firstinspires.ftc.twenty403.command.claw.ClawOpenCommand;
 import org.firstinspires.ftc.twenty403.command.drive.DriveCommand;
 import org.firstinspires.ftc.twenty403.command.drive.ResetGyroCommand;
 import org.firstinspires.ftc.twenty403.command.drive.SlowCommand;
 import org.firstinspires.ftc.twenty403.command.drive.TurboCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftDownCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftGroundJunctionCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftHighJunctionCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftIntakeCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftLowJunctionCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftMidJunctionCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftMoveDownOverrideCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftMoveUpOverrideCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftSetZeroCommand;
-import org.firstinspires.ftc.twenty403.command.lift.LiftUpCommand;
 import org.firstinspires.ftc.twenty403.helpers.BothButtons;
 
 import com.technototes.library.command.CommandScheduler;
-import com.technototes.library.command.ConditionalCommand;
 import com.technototes.library.control.CommandAxis;
 import com.technototes.library.control.CommandButton;
 import com.technototes.library.control.CommandGamepad;
 import com.technototes.library.control.Stick;
 
-public class Controls2 {
+public class ControlsDriver {
     public Robot robot;
     public CommandGamepad gamepad;
 
@@ -37,7 +24,7 @@ public class Controls2 {
     public CommandAxis driveStraight;
     public BothButtons override;
 
-    public Controls2(CommandGamepad g, Robot r) {
+    public ControlsDriver(CommandGamepad g, Robot r) {
         this.robot = r;
         gamepad = g;
         override = new BothButtons(g.leftTrigger.getAsButton(0.5));
@@ -47,12 +34,12 @@ public class Controls2 {
         if (Robot.RobotConstant.DRIVE_CONNECTED) {
             bindDriveControls();
         }
-        if (Robot.RobotConstant.CLAW_CONNECTED) {
-            bindClawControls();
-        }
-        if (Robot.RobotConstant.LIFT_CONNECTED) {
-            bindLiftControls();
-        }
+//        if (Robot.RobotConstant.CLAW_CONNECTED) {
+//            bindClawControls();
+//        }
+//        if (Robot.RobotConstant.LIFT_CONNECTED) {
+//            bindLiftControls();
+//        }
     }
 
     private void AssignNamedControllerButton() {
@@ -63,6 +50,7 @@ public class Controls2 {
         driveStraight = gamepad.rightTrigger;
 
         liftUpButton = gamepad.dpadRight;
+
         liftDownButton = gamepad.dpadDown;
         liftIntakePos = gamepad.dpadLeft;
 
@@ -90,29 +78,29 @@ public class Controls2 {
         resetGyroButton.whenPressed(new ResetGyroCommand(robot.drivebaseSubsystem));
     }
 
-    public void bindClawControls() {
-        // TODO: Name & Bind claw controls
-        clawOpenButton.whenPressed(new ClawOpenCommand(robot.clawSubsystem));
-        clawCloseButton.whenReleased(new ClawCloseCommand(robot.clawSubsystem));
-    }
-
-    public void bindLiftControls() {
-        // TODO: Name & Bind lift controls
-        liftUpButton.whenPressed(new LiftUpCommand(robot.liftSubsystem));
-        liftDownButton.whenPressed(new LiftDownCommand(robot.liftSubsystem));
-        liftIntakePos.whenPressed(new LiftIntakeCommand(robot.liftSubsystem));
-        liftOverrideZeroButton.whenPressed(
-                new ConditionalCommand(override, new LiftSetZeroCommand(robot.liftSubsystem)));
-
-        liftGroundOrOverrideDown.whenPressed(new ConditionalCommand(
-                override,
-                new LiftMoveDownOverrideCommand(robot.liftSubsystem),
-                new LiftGroundJunctionCommand(robot.liftSubsystem)));
-        liftLowOrOverrideUp.whenPressed(new ConditionalCommand(
-                override,
-                new LiftMoveUpOverrideCommand(robot.liftSubsystem),
-                new LiftLowJunctionCommand(robot.liftSubsystem)));
-        liftMedium.whenPressed(new LiftMidJunctionCommand(robot.liftSubsystem));
-        liftHigh.whenPressed(new LiftHighJunctionCommand(robot.liftSubsystem));
-    }
+//    public void bindClawControls() {
+//        // TODO: Name & Bind claw controls
+//        clawOpenButton.whenPressed(new ClawOpenCommand(robot.clawSubsystem));
+//        clawCloseButton.whenReleased(new ClawCloseCommand(robot.clawSubsystem));
+//    }
+//
+//    public void bindLiftControls() {
+//        // TODO: Name & Bind lift controls
+//        liftUpButton.whenPressed(new LiftUpCommand(robot.liftSubsystem));
+//        liftDownButton.whenPressed(new LiftDownCommand(robot.liftSubsystem));
+//        liftIntakePos.whenPressed(new LiftIntakeCommand(robot.liftSubsystem));
+//        liftOverrideZeroButton.whenPressed(
+//                new ConditionalCommand(override, new LiftSetZeroCommand(robot.liftSubsystem)));
+//
+//        liftGroundOrOverrideDown.whenPressed(new ConditionalCommand(
+//                override,
+//                new LiftMoveDownOverrideCommand(robot.liftSubsystem),
+//                new LiftGroundJunctionCommand(robot.liftSubsystem)));
+//        liftLowOrOverrideUp.whenPressed(new ConditionalCommand(
+//                override,
+//                new LiftMoveUpOverrideCommand(robot.liftSubsystem),
+//                new LiftLowJunctionCommand(robot.liftSubsystem)));
+//        liftMedium.whenPressed(new LiftMidJunctionCommand(robot.liftSubsystem));
+//        liftHigh.whenPressed(new LiftHighJunctionCommand(robot.liftSubsystem));
+//    }
 }
