+++ Twenty403/src/main/java/org/firstinspires/ftc/twenty403/opmode/ManualTeleOp2.java
@@ -1,6 +1,7 @@
 package org.firstinspires.ftc.twenty403.opmode;
 
-import org.firstinspires.ftc.twenty403.Controls.Controls2;
+import org.firstinspires.ftc.twenty403.Controls.ControlOperator;
+import org.firstinspires.ftc.twenty403.Controls.ControlsDriver;
 import org.firstinspires.ftc.twenty403.Hardware;
 import org.firstinspires.ftc.twenty403.Robot;
 import org.firstinspires.ftc.twenty403.command.autonomous.StartingPosition;
@@ -13,15 +14,15 @@ import com.technototes.library.util.Alliance;
 @TeleOp(name = "MeetNew TeleOp")
 public class ManualTeleOp2 extends CommandOpMode {
     public Robot robot;
-    public Controls2 controlsDriver;
-    public Controls2 controlsOperator;
+    public ControlsDriver controlsDriver;
+    public ControlOperator controlsOperator;
     public Hardware hardware;
 
     @Override
     public void uponInit() {
         hardware = new Hardware(hardwareMap);
         robot = new Robot(hardware, Alliance.NONE, StartingPosition.NEUTRAL);
-        controlsDriver = new Controls2(driverGamepad, robot);
-        controlsOperator = new Controls2(codriverGamepad, robot);
+        controlsDriver = new ControlsDriver(driverGamepad, robot);
+        controlsOperator = new ControlOperator(codriverGamepad, robot);
     }
 }
